apps/recommender/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
httpx==0.25.1
